+++ src/vue-timepicker.vue
@@ -45,6 +45,14 @@ export default {
     inputClass: { type: String },
     placeholder: { type: String },
     tabindex: { type: [ Number, String ], default: 0 },
+    inputWidth: { type: String },
+
+    hourLabel: { type: String },
+    minuteLabel: { type: String },
+    secondLabel: { type: String },
+    apmLabel: { type: String },
+    amText: { type: String },
+    pmText: { type: String },
 
     blurDelay: { type: [ Number, String ] },
     advancedKeyboard: { type: Boolean, default: false },
@@ -184,6 +192,13 @@ export default {
         formatString = formatString.replace(new RegExp(this.apmType, 'g'), this.apm)
       }
       return formatString
+    },
+
+    customDisplayTime () {
+      if (!this.amText && !this.pmText) {
+        return this.displayTime
+      }
+      return this.displayTime.replace(new RegExp(this.apm, 'g'), this.apmDisplayText(this.apm))
     },
 
     inputIsEmpty () {
@@ -362,6 +377,26 @@ export default {
         }
       }
       return range
+    },
+    
+    hourLabelText () {
+      return this.hourLabel || this.hourType
+    },
+    minuteLabelText () {
+      return this.minuteLabel || this.minuteType
+    },
+    secondLabelText() {
+      return this.secondLabel || this.secondType
+    },
+    apmLabelText () {
+      return this.apmLabel || this.apmType
+    },
+
+    inputWidthStyle () {
+      if (!this.inputWidth || !this.inputWidth.length) { return }
+      return {
+        width: this.inputWidth
+      }
     }
   },
 
@@ -892,6 +927,16 @@ export default {
       }
     },
 
+    apmDisplayText (apmValue) {
+      if (this.amText && (apmValue || '').toLowerCase() === 'am') {
+        return this.amText
+      }
+      if (this.pmText && (apmValue || '').toLowerCase() === 'pm') {
+        return this.pmText
+      }
+      return apmValue
+    },
+
     toggleDropdown () {
       if (this.disabled) { return }
       this.showDropdown = !this.showDropdown
@@ -1215,12 +1260,13 @@ export default {
 </script>
 
 <template>
-<span class="vue__time-picker time-picker">
+<span class="vue__time-picker time-picker" :style="inputWidthStyle">
   <input type="text" class="display-time"
          :class="[inputClass, {'disabled': disabled}]"
+         :style="inputWidthStyle"
          :id="id"
          :name="name"
-         :value="inputIsEmpty ? null : displayTime"
+         :value="inputIsEmpty ? null : customDisplayTime"
          :placeholder="placeholder || formatString"
          :tabindex="disabled ? null : tabindex"
          :disabled="disabled"
@@ -1230,12 +1276,12 @@ export default {
          @keydown.esc.exact="escBlur" />
   <span class="clear-btn" v-if="!showDropdown && showClearBtn" @click="clearTime">&times;</span>
   <div class="time-picker-overlay" v-if="showDropdown" @click="toggleDropdown"></div>
-  <div class="dropdown" v-show="showDropdown" @mouseup="keepFocusing" @click.stop="">
-    <div class="select-list">
+  <div class="dropdown" v-show="showDropdown" :style="inputWidthStyle" @mouseup="keepFocusing" @click.stop="">
+    <div class="select-list" :style="inputWidthStyle">
       <!-- Common Keyboard Support: less event listeners -->
       <template v-if="!advancedKeyboard">
         <ul class="hours">
-          <li class="hint" v-text="hourType"></li>
+          <li class="hint" v-text="hourLabelText"></li>
           <template v-for="(hr, hIndex) in hours">
             <li v-if="!opts.hideDisabledHours || (opts.hideDisabledHours && !isDisabledHour(hr))"
                 :key="hIndex"
@@ -1246,7 +1292,7 @@ export default {
           </template>
         </ul>
         <ul class="minutes">
-          <li class="hint" v-text="minuteType"></li>
+          <li class="hint" v-text="minuteLabelText"></li>
           <template v-for="(m, mIndex) in minutes">
             <li v-if="!opts.hideDisabledMinutes || (opts.hideDisabledMinutes && !isDisabledMinute(m))"
                 :key="mIndex"
@@ -1257,7 +1303,7 @@ export default {
           </template>
         </ul>
         <ul class="seconds" v-if="secondType">
-          <li class="hint" v-text="secondType"></li>
+          <li class="hint" v-text="secondLabelText"></li>
           <template v-for="(s, sIndex) in seconds">
             <li v-if="!opts.hideDisabledSeconds || (opts.hideDisabledSeconds && !isDisabledSecond(s))"
                 :key="sIndex"
@@ -1268,13 +1314,13 @@ export default {
           </template>
         </ul>
         <ul class="apms" v-if="apmType">
-          <li class="hint" v-text="apmType"></li>
+          <li class="hint" v-text="apmLabelText"></li>
           <template v-for="(a, aIndex) in apms">
             <li v-if="!opts.hideDisabledHours || (opts.hideDisabledHours && !isDisabledApm(a))"
                 :key="aIndex"
                 :class="{active: apm === a}"
                 :disabled="isDisabledApm(a)"
-                v-text="a"
+                v-text="apmDisplayText(a)"
                 @click="select('apm', a)"></li>
           </template>
         </ul>
@@ -1286,7 +1332,7 @@ export default {
       -->
       <template v-if="advancedKeyboard">
         <ul class="hours">
-          <li class="hint" v-text="hourType"></li>
+          <li class="hint" v-text="hourLabelText"></li>
           <template v-for="(hr, hIndex) in hours">
             <li v-if="!opts.hideDisabledHours || (opts.hideDisabledHours && !isDisabledHour(hr))"
                 :key="hIndex"
@@ -1308,7 +1354,7 @@ export default {
           </template>
         </ul>
         <ul class="minutes">
-          <li class="hint" v-text="minuteType"></li>
+          <li class="hint" v-text="minuteLabelText"></li>
           <template v-for="(m, mIndex) in minutes">
             <li v-if="!opts.hideDisabledMinutes || (opts.hideDisabledMinutes && !isDisabledMinute(m))"
                 :key="mIndex"
@@ -1330,7 +1376,7 @@ export default {
           </template>
         </ul>
         <ul class="seconds" v-if="secondType">
-          <li class="hint" v-text="secondType"></li>
+          <li class="hint" v-text="secondLabelText"></li>
           <template v-for="(s, sIndex) in seconds">
             <li v-if="!opts.hideDisabledSeconds || (opts.hideDisabledSeconds && !isDisabledSecond(s))"
                 :key="sIndex"
@@ -1352,7 +1398,7 @@ export default {
           </template>
         </ul>
         <ul class="apms" v-if="apmType">
-          <li class="hint" v-text="apmType"></li>
+          <li class="hint" v-text="apmLabelText"></li>
           <template v-for="(a, aIndex) in apms">
             <li v-if="!opts.hideDisabledHours || (opts.hideDisabledHours && !isDisabledApm(a))"
                 :key="aIndex"
@@ -1360,7 +1406,7 @@ export default {
                 :tabindex="isDisabledApm(a) ? null : tabindex"
                 :data-key="a"
                 :disabled="isDisabledApm(a)"
-                v-text="a"
+                v-text="apmDisplayText(a)"
                 @click="select('apm', a)"
                 @keydown.space.prevent="select('apm', a)"
                 @keydown.enter.prevent="select('apm', a)"